Refactoring (after review)
