+++ src/VirtoCommerce.Platform.Data/GenericCrud/SearchService.cs
@@ -1,7 +1,6 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Caching.Memory;
@@ -20,20 +19,12 @@ namespace VirtoCommerce.Platform.Data.GenericCrud
     /// Generic service to simplify search implementation.
     /// To implement the service for applied purpose, inherit your search service from this.
     /// <para>
-    /// This base provides the <see cref="IExtendedSearchService{TCriteria, TResult, TModel}"/> members
-    /// (<see cref="SearchAllAsync"/>, <see cref="SearchAllIdsAsync"/>, <see cref="CountAllAsync"/>, <see cref="ExistsAsync"/>)
-    /// but does NOT implement the interface itself, so the optimized "search all" path is opt-in.
-    /// <see cref="SearchAllAsync"/> reads the identifiers of all matching data with a single query
-    /// (no count query, no offset paging) and then loads the models by identifiers in batches
-    /// of <see cref="CrudOptions.SearchAllBatchSize"/>. Compared to Skip/Take paging via
-    /// <see cref="SearchAsync"/>, this avoids re-evaluating the search query and counting the records
-    /// for every batch, and makes the returned data immune to page drift caused by concurrent modifications.
-    /// It follows that these members are correct only when the result membership is fully defined by
-    /// <see cref="BuildQuery"/>/<see cref="GetOrderedQueryAsync"/>. A service that meets this condition
-    /// enables the optimized path by deriving from <see cref="ExtendedSearchService{TCriteria, TResult, TModel, TEntity}"/>
-    /// or by declaring <see cref="IExtendedSearchService{TCriteria, TResult, TModel}"/> on the service itself;
-    /// a service whose <see cref="SearchAsync"/> changes the result membership must additionally override
-    /// <see cref="SearchAllAsync"/>, <see cref="SearchAllIdsAsync"/>, <see cref="CountAllAsync"/> and <see cref="ExistsAsync"/>.
+    /// To additionally expose the optimized "search all" operations
+    /// (<see cref="IExtendedSearchService{TCriteria, TResult, TModel}"/> — a single identifiers query without
+    /// per-page counting, then loading models by identifiers in batches), derive from
+    /// <see cref="ExtendedSearchService{TCriteria, TResult, TModel, TEntity}"/> instead of this class.
+    /// Do so only when the result membership is fully defined by <see cref="BuildQuery"/>/<see cref="GetOrderedQueryAsync"/>
+    /// (i.e. <see cref="SearchAsync"/> is not overridden in a way that changes which records are returned).
     /// </para>
     /// </summary>
     /// <typeparam name="TCriteria">Search criteria type (a descendant of <see cref="SearchCriteriaBase"/>)</typeparam>
@@ -95,7 +86,7 @@ namespace VirtoCommerce.Platform.Data.GenericCrud
         /// Builds a <typeparamref name="TResult"/> for the given ordered identifiers: loads the models,
         /// restores the identifier order (<see cref="ICrudService{TModel}.GetAsync"/> returns them in cache order),
         /// and runs <see cref="ProcessSearchResultAsync"/>. Shared by <see cref="SearchAsync"/> (one page) and
-        /// <see cref="SearchAllAsync"/> (one batch) so both assemble and post-process results identically.
+        /// the optimized "search all" path (one batch) so both assemble and post-process results identically.
         /// </summary>
         protected virtual async Task<TResult> CreateSearchResultAsync(IList<string> ids, int totalCount, TCriteria criteria, bool clone)
         {
@@ -109,102 +100,6 @@ namespace VirtoCommerce.Platform.Data.GenericCrud
             }
 
             return await ProcessSearchResultAsync(result, criteria);
-        }
-
-        /// <inheritdoc cref="IExtendedSearchService{TCriteria, TResult, TModel}.SearchAllAsync" />
-        public virtual async Task<IList<TModel>> SearchAllAsync(TCriteria criteria, bool clone = true, CancellationToken cancellationToken = default)
-        {
-            // Skip/Take are ignored by SearchAll*; normalize them so paging never leaks into the
-            // result assembly (e.g. a ProcessSearchResultAsync override that slices by criteria.Skip/Take).
-            criteria = WithoutPaging(criteria);
-
-            var ids = await SearchAllIdsAsync(criteria, cancellationToken);
-
-            var result = new List<TModel>();
-
-            foreach (var batchIds in ids.Paginate(_crudOptions.SearchAllBatchSize))
-            {
-                cancellationToken.ThrowIfCancellationRequested();
-
-                // Assemble each batch through the same pipeline as a regular search page. Running
-                // ProcessSearchResultAsync per batch (rather than once over the whole set) keeps any
-                // result-dependent enrichment inside it bounded to SearchAllBatchSize items.
-                var batchResult = await CreateSearchResultAsync(batchIds, ids.Count, criteria, clone);
-
-                result.AddRange(batchResult.Results);
-            }
-
-            return result;
-        }
-
-        /// <inheritdoc cref="IExtendedSearchService{TCriteria, TResult, TModel}.SearchAllIdsAsync" />
-        public virtual Task<IList<string>> SearchAllIdsAsync(TCriteria criteria, CancellationToken cancellationToken = default)
-        {
-            criteria = WithoutPaging(criteria);
-            ValidateSearchCriteria(criteria);
-
-            return GetOrCreateCachedAsync(nameof(SearchAllIdsAsync), criteria,
-                ct => SearchAllIdsNoCacheAsync(criteria, ct), cancellationToken);
-        }
-
-        /// <inheritdoc cref="IExtendedSearchService{TCriteria, TResult, TModel}.CountAllAsync" />
-        public virtual Task<int> CountAllAsync(TCriteria criteria, CancellationToken cancellationToken = default)
-        {
-            criteria = WithoutPaging(criteria);
-            ValidateSearchCriteria(criteria);
-
-            return GetOrCreateCachedAsync(nameof(CountAllAsync), criteria, async ct =>
-            {
-                using var repository = _repositoryFactory();
-                return await BuildQuery(repository, criteria).CountAsync(ct);
-            }, cancellationToken);
-        }
-
-        /// <inheritdoc cref="IExtendedSearchService{TCriteria, TResult, TModel}.ExistsAsync" />
-        public virtual Task<bool> ExistsAsync(TCriteria criteria, CancellationToken cancellationToken = default)
-        {
-            criteria = WithoutPaging(criteria);
-            ValidateSearchCriteria(criteria);
-
-            return GetOrCreateCachedAsync(nameof(ExistsAsync), criteria, async ct =>
-            {
-                using var repository = _repositoryFactory();
-                return await BuildQuery(repository, criteria).AnyAsync(ct);
-            }, cancellationToken);
-        }
-
-        /// <summary>
-        /// Returns a copy of the criteria with paging reset (Skip = Take = 0). SearchAll* operations
-        /// return every matching record, so paging must not affect their results or cache keys.
-        /// The original instance is returned unchanged when it already carries no paging.
-        /// </summary>
-        private static TCriteria WithoutPaging(TCriteria criteria)
-        {
-            if (criteria.Skip == 0 && criteria.Take == 0)
-            {
-                return criteria;
-            }
-
-            criteria = criteria.CloneTyped();
-            criteria.Skip = 0;
-            criteria.Take = 0;
-
-            return criteria;
-        }
-
-        /// <summary>
-        /// Wraps a query result in the shared exclusive-cache scaffolding used by the SearchAll* operations
-        /// (cache key from the operation name and criteria, plus the standard expiration token).
-        /// </summary>
-        private Task<T> GetOrCreateCachedAsync<T>(string operationName, TCriteria criteria, Func<CancellationToken, Task<T>> factory, CancellationToken cancellationToken)
-        {
-            var cacheKey = CacheKey.With(GetType(), operationName, criteria.GetCacheKey());
-
-            return _platformMemoryCache.GetOrCreateExclusiveAsync(cacheKey, cacheOptions =>
-            {
-                cacheOptions.AddExpirationToken(CreateCacheToken(criteria));
-                return factory(cancellationToken);
-            });
         }
 
         protected virtual void ValidateSearchCriteria(TCriteria criteria)
@@ -258,18 +153,6 @@ namespace VirtoCommerce.Platform.Data.GenericCrud
             }
 
             return result;
-        }
-
-        protected virtual async Task<IList<string>> SearchAllIdsNoCacheAsync(TCriteria criteria, CancellationToken cancellationToken)
-        {
-            using var repository = _repositoryFactory();
-
-            var query = BuildQuery(repository, criteria);
-            var orderedQuery = await GetOrderedQueryAsync(query, criteria);
-
-            return await orderedQuery
-                .Select(x => x.Id)
-                .ToListAsync(cancellationToken);
         }
 
         protected virtual Task<IOrderedQueryable<TEntity>> GetOrderedQueryAsync(IQueryable<TEntity> query, TCriteria criteria)
